Minimal WordPress template for use as a wrapper in other CMS', such as Joomla.
This template is used for embedding a WordPress web site in another web
site through a wrapper or iframe. It only displays the content of the
site and a right sidebar. No header or footer information is displayed.
It was designed to take as little space as possible to fit within a
frame of another web site.